함수 역할에 따라 관심사 분리
