+++ src/refactoring/models/coupon.ts
@@ -1,8 +1,15 @@
 import { Coupon } from "../../types";
 
-export const createNewCoupon = (): Coupon => ({
-  name: "",
-  code: "",
-  discountType: "percentage",
-  discountValue: 0,
-});
+export const applyCouponDiscount = (
+  totalAmount: number,
+  selectedCoupon: Coupon
+) => {
+  switch (selectedCoupon.discountType) {
+    case "amount":
+      return totalAmount - selectedCoupon.discountValue;
+    case "percentage":
+      return totalAmount * (1 - selectedCoupon.discountValue / 100);
+    default:
+      return totalAmount;
+  }
+};
